requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
prometheus_client==0.21.0
requests==2.32.3
schedule==1.2.2
speedtest-cli==2.1.3
urllib3==2.2.3
